+++ packages/multi-tenant/src/lib/getDatabaseConfig.ts
@@ -1,14 +1,23 @@
 import type { SlonikConfig } from "@dzangolab/fastify-slonik";
-import type { MigrateDBConfig } from "@dzangolab/postgres-migrations";
+import type { ClientConfig } from "pg";
 
-const getDatabaseConfig = (slonikConfig: SlonikConfig): MigrateDBConfig => {
-  return {
+const getDatabaseConfig = (slonikConfig: SlonikConfig): ClientConfig => {
+  let clientConfig: ClientConfig = {
     database: slonikConfig.db.databaseName,
     user: slonikConfig.db.username,
     password: slonikConfig.db.password,
     host: slonikConfig.db.host,
     port: slonikConfig.db.port,
-  } as MigrateDBConfig;
+  };
+
+  if (slonikConfig.clientConfiguration?.ssl) {
+    clientConfig = {
+      ...clientConfig,
+      ssl: slonikConfig.clientConfiguration?.ssl,
+    };
+  }
+
+  return clientConfig;
 };
 
 export default getDatabaseConfig;