+++ packages/s3/src/index.ts
@@ -1,10 +1,14 @@
+import "@dzangolab/fastify-mercurius";
+import type { FilenameResolutionStrategy } from "./types";
+
 declare module "@dzangolab/fastify-config" {
   interface ApiConfig {
     s3: {
       accessKey: string;
       bucket: string | Record<string, string>;
       endPoint?: string;
-      fileSizeLimit?: number;
+      fileSizeLimitInBytes?: number;
+      filenameResolutionStrategy?: FilenameResolutionStrategy;
       forcePathStyle?: boolean;
       secretKey: string;
       region?: string;
@@ -19,5 +23,7 @@ export { default as FileService } from "./model/files/service";
 export { default as S3Client } from "./utils/s3Client";
 export type { FilePayload, Multipart } from "./types";
 export type { File, FileCreateInput, FileUpdateInput } from "./types/file";
+export type { FileUpload as GraphQLFileUpload } from "graphql-upload-minimal";
 
 export { default } from "./plugin";
+export { default as multipartParserPlugin } from "./plugins/multipartParser";