+++ packages/user/src/model/users/handlers/index.ts
@@ -1,6 +1,8 @@
 import adminSignUp from "./adminSignUp";
 import canAdminSignUp from "./canAdminSignUp";
 import changePassword from "./changePassword";
+import disable from "./disable";
+import enable from "./enable";
 import me from "./me";
 import updateMe from "./updateMe";
 import users from "./users";
@@ -9,6 +11,8 @@ export default {
   adminSignUp,
   canAdminSignUp,
   changePassword,
+  disable,
+  enable,
   me,
   updateMe,
   users,